Field class derives from ModArith to support rings
